pyproject.toml
[tool.poetry]
name = "livraria"
version = "0.1.0"
description = "Livraria Legal"
authors = ["also"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.13.0"
gunicorn = "^20.1.0"
whitenoise = "^6.2.0"
djangorestframework-simplejwt = "^5.2.1"
python-magic = "^0.4.27"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]
black = {version = "^22.8.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
